feat/updated AgentList and ConnectionsProtocol.js for OOB support
