+++ src/renderer/components/AgentList.vue
@@ -83,10 +83,12 @@ const electron = require('electron');
 import { mapState, mapActions, mapGetters } from "vuex"
 import message_bus from '@/message_bus.js';
 import { from_store } from '../connection_detail.js';
+import { base64_decode } from '../base64.js';
 import ConnectionsProtocol from './ConnectionsProtocol.js';
 const uuidv4 = require('uuid/v4');
 const coordinate_mediation =
   (type) => `did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/coordinate-mediation/1.0/${type}`;
+
 export default {
   name: 'agent-list',
   components: {  },
@@ -122,6 +124,7 @@ export default {
   },
   methods: {
     ...mapActions("Agents", ["add_agent", "update_agent", "delete_agent"]),
+
     openConnection: async function(a) {
       const modalPath = process.env.NODE_ENV === 'development'
         ? 'http://localhost:9080/#/agent/'+a.id
@@ -136,6 +139,7 @@ export default {
           enableRemoteModule: true,
         }
       })
+
       // TODO: Get the close handler to work to keep track of open windows only
       win.on('close', function () {
         win = null;
@@ -144,10 +148,13 @@ export default {
       win.loadURL(modalPath);
       let window_key = a.my_key_b58.publicKey;
       this.agent_windows[window_key] = win;
+
     },
+
     deleteConnection: async function(a){
       this.delete_agent(a);
     },
+
     mediatorConnect: async function(mediator_agent){
       let vm = this; //hold reference
       let agent_info = await this.get_agent(mediator_agent.id);
@@ -171,6 +178,7 @@ export default {
       });
       console.log("mediator connected", this.mediator_connection);
     },
+
     /**
      * Process messages received through mediator.
      */
@@ -194,6 +202,7 @@ export default {
         // TODO: get inbound message to proper agent window. this needs to happen even if window not open.
         // store message
         let message_uuid = uuidv4();
+
         // send message
         let window_key = recipients[0];
         // TODO: detect lack of open window.
@@ -217,6 +226,7 @@ export default {
         //TODO: deliver queued message after window opens.
       };
     },
+
     /**
      * Tear down mediator info after mediator connection deletion.
      */
@@ -230,12 +240,14 @@ export default {
       }
       this.mediator_connection = null;
     },
+
     /**
      * Generate an invitation for connecting to the toolbox through mediation.
      */
     generate_invitation: function(){
       // TODO Implement
     },
+
     /**
      * Inform the mediator of a new key to route to this agent.
      * @param {string} verkey Key to add to mediator routes.
@@ -260,6 +272,7 @@ export default {
         // TODO do something with errors
       }
     },
+
     async send_mediation_request(connection) {
       // prepare message
       let msg = {
@@ -281,6 +294,7 @@ export default {
         console.error("Encountered error while attempting to establish mediation:", err)
       }
     },
+
     async process_mediator_invitation() {
       let connection = await this.new_agent_invitation_process(this, this.new_mediator_invitation);
       connection.unpacked_processor = this.mediatorInbound(connection);
@@ -288,7 +302,21 @@ export default {
       this.mediatorConnect(connection);
       this.new_mediator_invitation = "";
     },
+
+    getUrlVars: function(url) {
+        var vars = {};
+        var parts = url.replace(/[?&]+([^=&]+)=([^&]*)/gi, function(m,key,value) {
+          vars[key] = value;
+        });
+        return vars;
+      },
+
     async connect_clicked() {
+      var invite_b64 = this.getUrlVars(this.new_agent_invitation)["c_i"];
+      //base 64 decode
+      var invite_string = base64_decode(invite_b64);
+      var invite = JSON.parse(invite_string);
+      
       this.invitation_error = "";
       try {
         await ConnectionsProtocol.new_agent_invitation_process(this, this.new_agent_invitation);
@@ -320,6 +348,7 @@ export default {
       mediator_connection: {},
       window_message_queue: {},
     }
+
   }
 }
 </script>
